+++ src/Blockcore.Explorer/ClientApp/src/environments/environment.prod.ts
@@ -1,4 +1,5 @@
 export const environment = {
   production: true,
   useLocalIndexer: false,
+  local: false // Set this to true to load your local "chain.json".
 };